Correct the name of the files
